[io] Record test_sequence's check floors; fix its one 32-bit format string
The floor rows are the sanctioned follow-up to adding a suite: verify.sh
--update-checks-baseline, run green across all four configurations (summary:
opencv 2436721 checks, core 1866875, noexcept 1866798+60s, debug 1866865, all
PASS), recorded test_sequence at 73 checks and 0 skips in each. The counts are
equal across configurations because the suite needs neither OpenCV nor
exceptions -- the point of the reader it verifies.

The printf cast is verify_cortex_m.sh earning its keep before the commit ever
reached it: uint32_t is `unsigned long` on arm-none-eabi, so `%u` against
SequenceHeader::mode was a -Wformat error at 32-bit size_t and invisible to
every x86 configuration. With the cast, all 34 suites -- test_sequence now
among them -- compile clean for Cortex-M7.

+++ tests/test_sequence.cpp
@@ -145,7 +145,8 @@ BINCV_TEST(Sequence, HeaderParsesAndRejects) {
     const std::vector<uint8_t> h8 = blobHeader(0, 32, 24, 3);
     const SequenceHeader a = readSequenceHeader(h8.data(), h8.size());
     std::printf(" 8-bit header: valid=%d mode=%u %zux%zu x%zu, %zu B/frame\n",
-                a.valid ? 1 : 0, a.mode, a.width, a.height, a.frameCount, a.frameBytes);
+                a.valid ? 1 : 0, static_cast<unsigned>(a.mode), a.width, a.height,
+                a.frameCount, a.frameBytes);
     BINCV_CHECK(a.valid);
     BINCV_CHECK(a.mode == kSequenceMode8Bit);
     BINCV_CHECK(a.width == 32 && a.height == 24 && a.frameCount == 3);
